requirements.txt
fastapi
uvicorn
requests
langchain
langchain-community
langchain-openai
python-dotenv
networkx
unstructured